requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
shap==0.46.0
joblib==1.4.2
pytest==8.3.4
httpx==0.28.1

# Hybrid C++ core
pybind11==2.13.6

# Redis (rate limits, drift stats)
redis==5.0.8

# OpenTelemetry
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-exporter-otlp==1.28.2
